requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asyncio==3.4.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.5
jiter==0.8.2
kiwisolver==1.4.8
lingua-language-detector==2.0.2
llvmlite==0.44.0
lv-linkpy==1.0.3
MarkupSafe==3.0.2
matplotlib==3.10.1
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai==1.65.1
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
pillow==10.4.0
proglog==0.1.10
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
regex==2024.11.6
requests==2.32.3
scipy==1.15.2
six==1.17.0
slack_sdk==3.34.0
sniffio==1.3.1
SPMUtil==1.0.5
sympy==1.13.3
tiktoken==0.9.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
